requirements.txt
gradio
python-dotenv
langchain
langchain-openai
pyvis
PyMuPDF
flask
networkx